voxtype/requirements.txt
PySide6>=6.8.0
aiohttp>=3.9
pynput>=1.7.6
pyautogui>=0.9.54
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26
Pillow>=10.2
mss>=9.0
pywin32>=306; platform_system == "Windows"
